Создали таблицу для подтверждения почты, связли с моделью user. а user с таблицей кодов, т.е. создали модель. Также по умолчанию создалось Police как и с аккаунтом(хз для чего) ну и сидры и фэктори для тестов. Теперь можно приступать к роутам редактирвоания.

+++ database/migrations/2024_04_06_151036_create_accounts_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('accounts', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id'); // ID пользователя из таблицы Users
+            $table->integer('user_id'); // ID пользователя из таблицы Users
             $table->string('account_source_type'); // Источник учетной записи (URL, app, etc.)
             $table->string('username'); // Логин
             $table->string('password'); // Пароль
